plugin/py/requirements.txt
argparse
ecdsa
fastecdsa
sympy
cairo-lang
pytest
python-dotenv
